insurance/requirements.txt
Django==2.2.6
psycopg2==2.7.7
djangorestframework==3.10.3
black==19.3b0
